+++ src/impls/alloc_impls.rs
@@ -10,9 +10,11 @@ use alloc::vec::Vec;
 
 use itertools::Itertools as _;
 
-use crate::iteration::peekable_exhaust;
+use crate::iteration::{peekable_exhaust, Pei};
 use crate::patterns::{delegate_factory_and_iter, impl_newtype_generic};
 use crate::Exhaust;
+
+// -------------------------------------------------------------------------------------------------
 
 impl_newtype_generic!(T: [], Box<T>, Box::new);
 impl_newtype_generic!(T: [], Rc<T>, Rc::new);
@@ -43,26 +45,93 @@ impl<T: Exhaust + Ord> Exhaust for BTreeSet<T> {
     }
 }
 
-// TODO: Instead of delegating to itertools, we should implement our own powerset
-// iterator, to provide the preferred ordering of elements.
-pub struct ExhaustSet<T: Exhaust>(itertools::Powerset<<T as Exhaust>::Iter>);
+// -------------------------------------------------------------------------------------------------
+
+/// Iterator which exhausts set types, in sorted order if `T::exhaust()` is sorted.
+pub struct ExhaustSet<T: Exhaust> {
+    /// The next elements of these iterators are all but one of the elements of the next set to be
+    /// produced (unless `self.last` is exhausted, in which case we will start backtracking).
+    ///
+    /// Exception: In the case of the empty set, this is empty, since it can't have -1 elements
+    /// as the above rule would otherwise imply.
+    list: Vec<Pei<T>>,
+
+    /// If `None`, we should produce the empty set (as the first element).
+    /// Otherwise, contains an iterator whose next item is the greatest element of `T` that the
+    /// next set produced whose other elements are in `list` will contain.
+    ///
+    /// This iterator is exhausted when we are backtracking (in which case `self.list` is
+    /// non-empty) or when we have produced every set (in which case `self.list` is empty).
+    last: Option<Pei<T>>,
+}
 
 impl<T: Exhaust> Default for ExhaustSet<T> {
     fn default() -> Self {
-        ExhaustSet(itertools::Itertools::powerset(T::exhaust_factories()))
+        ExhaustSet {
+            list: Vec::new(),
+            last: None,
+        }
     }
 }
 
 impl<T: Exhaust> Clone for ExhaustSet<T> {
     fn clone(&self) -> Self {
-        Self(self.0.clone())
+        Self {
+            list: self.list.clone(),
+            last: self.last.clone(),
+        }
     }
 }
 
 impl<T: Exhaust> Iterator for ExhaustSet<T> {
     type Item = Vec<T::Factory>;
     fn next(&mut self) -> Option<Self::Item> {
-        self.0.next()
+        let Some(ref mut last_iter) = self.last else {
+            // Initial state; time to produce the set with 0 elements, and
+            // set up to produce a set with 1 element after this one.
+            self.last = Some(peekable_exhaust::<T>());
+            return Some(Vec::new());
+        };
+
+        loop {
+            return if let Some(elem) = last_iter.peek() {
+                let item = Vec::from_iter(
+                    self.list
+                        .iter_mut()
+                        .map(|iter| {
+                            iter.peek()
+                                .expect("no peekable should be exhausted")
+                                .clone()
+                        })
+                        .chain(iter::once(elem.clone())),
+                );
+
+                // Start iterating over sets that are 1 element longer than this one,
+                // if there are any.
+                self.list.push(last_iter.clone());
+
+                // Now that we’ve grabbed the clone we need for self.list, actually advance
+                // the iterator in self.last (we only peeked before).
+                last_iter.next();
+
+                Some(item)
+            } else {
+                // self.last is exhausted.
+                // This means there are no more sets of length self.list.len() + 1
+                // with the prefix currently stored in self.list.
+                // Advance the iterator in self.list and make it the new self.last.
+                if let Some(mut new_last) = self.list.pop() {
+                    _ = new_last.next();
+                    *last_iter = new_last;
+                    // TODO: can we express this algorithm without a loop?
+                    continue;
+                } else {
+                    // No more sets at all.
+                    // Keep the exhausted iterator as a marker to stop (so we're fused).
+                    None
+                }
+            };
+        }
     }
 }
 impl<T: Exhaust> FusedIterator for ExhaustSet<T> {}
@@ -73,9 +142,15 @@ where
     T::Iter: fmt::Debug,
 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_tuple("ExhaustSet").field(&self.0).finish()
+        let Self { list, last } = self;
+        f.debug_struct("ExhaustSet")
+            .field("list", &list)
+            .field("last", &last)
+            .finish()
     }
 }
+
+// -------------------------------------------------------------------------------------------------
 
 pub(crate) type MapFactory<K, V> = Vec<(<K as Exhaust>::Factory, <V as Exhaust>::Factory)>;